Bugfix: Replicate textage scrlist.js logic for difficulty matching
This fixes our logic to match textages such that the listings
between INFINITAS only songs and current version AC songs
match the site.

Unfortunately, as of 2025-10-30, there are bugs in their data!
An example is that 'Flash Back 90s LEGGENDARIA' is available
in INFINITAS, but does not have its chart listed when linked
from the infinitas only page.

We are considering sourcing our data elsewhere, or normalizing
this using Konami's public site for INFINITAS songs.